drivers: add stm32 RIFSC support
Add the RIFSC new driver support.

RIFSC (RIF Security Controller) is responsible for the isolation
of hardware resources like memory or peripherals. It is composed of:

-RISC registers(slave peripherals) with RISUP(Resource Isolation
Slave Unit for Peripherals) OR RISAL(Resource Isolation Slave Unit
for Address space - Lite) logics.
-RIMC registers(Non RIF-Aware masters counterpart) with RIMU
(Resource Isolation Master Unit) logic. It is possible for a master to
inherit from its slave port(RISUP) configuration.

This driver parses the RIFSC device tree configuration and applies
it to put the firewall in place. Therefore, the device tree is
mandatory.

Change-Id: I2e320ad9f83aa945e8232158c982e386e78ccf32
Signed-off-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Signed-off-by: Ludovic Barre <ludovic.barre@foss.st.com>
Signed-off-by: Christophe Kerello <christophe.kerello@foss.st.com>