drivers: stm32mp25: new FMC driver with RIF support
This driver implements RIF configuration for FMC, which is a RIF aware IP.
It means that the FMC driver is in charge of configuring its own RIF
restrictions and that the FMC has dedicated RIF configuration registers.

Additional check on RIF configuration is added for this IP when debug is
on.

Signed-off-by: Gatien Chevallier <gatien.chevallier@st.com>
Change-Id: Id4052517e87d058813dd8d7a3cfd24a3e5c7d793
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/242559
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>