clk: stm32mp13: add adc bus clocks
Bus clock is mandatory as also needed for register interface.
'adc' clock is optional: it's the one used for analog when specified
in the device tree.

Current clock driver for MP13 exposes only ADC kernel clock.
There's an issue then for clock gating when only the BUS clock
is used for both registers access and analog.
Then we have to manage a counter at the gate level.

Change-Id: I70556a98c3de24a7a509912b1dae0600291493fb
Signed-off-by: Gabriel Fernandez <gabriel.fernandez@foss.st.com>