drivers: stm32_stgen: add STGEN support
STGEN final configuration is under OP-TEE responsibility because
it updates clock configuration.

This patch implements the support of the STGEN configuration,
which should be done in a driver. The system base counter
(STGEN counter) should be based on HSE clock. Because OP-TEE
reconfigures the clock tree, it is OP-TEE's responsibility to ensure
this.

ARM CP-15 registers must be updated according to system clock frequency.
Only the highest EL (In our case the secure monitor) can do the
CP15 update. So, this patch implements this mechanism with a SMC
call.

Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Change-Id: I60156de655fe8bf673129262711d17f8c9d4e5c1
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/261701
Tested-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Domain-Review: Lionel DEBIEVE <lionel.debieve@foss.st.com>