drivers: stm32_stgen: support low-power sequence
In some low-power modes such as standby, the STGEN context is lost and
the STGEN is powered off.

In order to correctly restore the platform counter, it is necessary
to save the counter value and the current time from the RTC before
disabling the STGEN.

When resuming the platform, the STGEN should be restored with the
correct clock configuration. The time spent in low-power is deduced
from the time read from the RTC. The STGEN counter value is then
updated with the corresponding number of ticks + the counter value
before the suspend.

Change-Id: Ibeb51fd1cc2961122926207ba9df7fe66d1ee81f
Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/319320
Tested-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>