drivers: stm32_rifsc: add semaphore handling in low-power sequences
On STM32MP2x platforms, when coming back from a low-power mode where
the context is lost, the RIF semaphores are released by the ROM code
but the RIFSC configuration is kept.

It is necessary to save which RIFSC semaphore were taken before
going to such low-power modes and to restore them when resuming.

Change-Id: Id9aa15a1658838de25242cdea9e0955be7fea84f
Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>