core: dt_driver: don't enforce phandle 1st arg is a phandle
Changes local function device_from_provider_prop() to assume its
argument @prop points to the first argument to pass with phandle.

This change allows a later change to support other DT bindings
("interrupts" property) where 1st cell of the property is not
a phandle but the 1st phandle argument to be passed.

Change-Id: Iabbcc1e0e39379ed33115a3b57618c0e28b8ee37
[etienne: cherry-picked from 955b02aaf75cc8fdc66006ba3daf6c19c3ded476]
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>