plat-stm32mp1: default disable reserved shared memory
Changes stm32mp15 default configuration to not embed OP-TEE's reserved
static shared memory. This change is aligned with a recent change in
TF-A [1] that also default disables the related configuration switch.
Note that TF-A/stm32mp1 deprecates this configuration in order to fully
remove its support in the future [2].

Note that when CFG_CORE_RESERVED_SHM is disabled the default 2MB area
is now used by OP-TEE as secure memory.

Change-Id: I3f3e945f2fa2e25e246c6be62686dbfa0ec8e0f1
Link: https://git.trustedfirmware.org/TF-A/trusted-firmware-a.git/commit/?id=fb1d3bd9330ce70f735a344dd4223faffb261118 [1]
Link: https://git.trustedfirmware.org/TF-A/trusted-firmware-a.git/commit/?id=12e683a68049f6a3d0985a2cd1564e00115e809f [2]
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>
[etienne: cherry-picked from 8b3ac1f6bc0df62a6524bed9f21a4c97bc8fe3ce]
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>