core: arm: check for NULL mobj before thread_rpc_free()
In the SMC and FF-A ABIs check that a mobj representing a shared memory
object isn't NULL before doing an RPC to free it in the normal world
too. For the FF-A ABI it's harmless without this check, but the SMC ABI
may cause an NULL pointer dereference in the OP-TEE kernel driver.

Change-Id: I4dca71f037ba5b434fae5636102bb8dad9ea272d
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
[etienne: cherry-picked from 4989730f373a199c910f992e2a22325615241b81]
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>