drivers: regulator: fix getting GPIO for level control
Swap the 2 calls to find gpio property in "regulator-gpio" compatible
node. The call that expects to get a TEE_ERROR_ITEM_NOT_FOUND return
value (as described in the inline comment) for an unsupported 2nd level
control GPIO can fail with TEE_ERROR_DEFER_DRIVER_INIT return value
when the 1st GPIO phandle listed in the 'gpios' DT node property relates
to a GPIO driver that is not yet probed. Fix that by first getting the
required GPIO control level and then testing presence of other GPIO
phandle in the 'gpios' DT property.

Change-Id: I5107460e4bc5656bc2bbbe7ae23a1b63634a3388
Fixes: f164f0f83420 ("drivers: regulator: GPIO controlled regulator")
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>
[etienne: cherry-picked from 950549e402556186ed1fc2632afeb6c1c63b2839]
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>