drivers: clk: remove spinlock in clk_print_tree()
Remove spinlock handling from clk_print_tree(). This lock may mask
interrupts for a long period of time which and may affect the system
behavior. Since there is no clock unregistering and clock destroying
function, all clock references used are safe to use, even outside
the locked state. Whether or not printed information is accurate is
not an issue as clk_print_tree() is concerned as the function is only
informative.

Change-Id: I18d9e128e7b0020965ad66ed420b2bdf9cc8c7a2
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>
[etienne: cherry-picked from 4362d4f975f6816bcdcd8cb52de4165db2b1cd17]
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>