core: fix objectUsage handling for persistent objects
Before this patch was the object usage field stored in the
non-persistent part of an object handle, regardless of whether the
storage object was persistent. This prevents updates to this field from
being restored the next time the persistent object is opened. Updates to
the field are also not replicated to eventual other open handles for the
object. Fix this by storing the "usage" bits in a new obj_info_usage
field in struct tee_pobj for persistent objects. Updates to the field
are also written into secure storage to preserve the content the next
time the object is opened.

Change-Id: I12eafa9f02137c5cdc82c2be74148480a03cd2e1
Fixes: b01047730e77 ("Open-source the TEE Core")
Closes: https://github.com/OP-TEE/optee_os/issues/6495
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
[etienne: cherry-picked from 0f50ba5ade6decc49209304b103cf208c321ade6]
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>