core: panic(): avoid panic in panic
When configured with CFG_HALT_CORES_ON_PANIC=y panic() sends a secure
SGI to all other CPUs to trap all of them, but if interrupts haven't
been configured yet it triggers a cascade of panics. So add a check to
see if interrupts are configured before trying to trap the other CPUs.

Change-Id: Ibfac00031aa7d3529691229f811ecad1bb23ac40
Fixes: 2b719df0fc1c ("core: interrupt: halt other cores when one is panicking")
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
[etienne: cherry-picked from 01ef3c7903e8a54f24a9245feca4644941b6bc62]
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>