drivers: stm32_rstctrl: move stm32mp1x controller in stm32mp1_rstcrl.c
This patch introduces the STM32MP25 reset controller.
The binding for the STM32MP25 is different from that of the
STM32MP1x, so we need to create a stm32mp25_rstcrl.c file.
We will keep a generic API stm32_rstcrl.c file for probing and
passing platform data.

Change-Id: Ie5c72e1ceb55d3e597c938c502fce434f4007b16
Signed-off-by: Gabriel Fernandez <gabriel.fernandez@foss.st.com>