drivers: stm32_gpio: add secure configuration for GPIOs
This change adds security support for GPIOS. A bank of GPIO now has a
secure support and configuration.

Secure support is defined in the device tree. If a GPIO bank is defined
as secure, the secure configuration is read through st,protreg device
tree property and is applied during probe.

Change-Id: I345b4274b2ffa154094fe617c5f10505d0d0fc3a
Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>