dts: stm32: add wakeup properties to the RTC
The 'wakeup-source' property allows to say that the RTC peripheral
can wakeup the platform.
Overload the property 'wakeup-parent' for the RTC as the wakeup-parent
capable of wakeup the platform with a RTC signal input is the exti2.

Change-Id: I81e4566e165dc61a4efaeab28d95aca18a31cf18
Signed-off-by: Clément Le Goffic <clement.legoffic@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/402363
Domain-Review: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>