drivers: stm32_rtc: add the capability to wakeup the platform
During probe :
- We now look for the property "wakeup-source" and "wakeup-parent".
"wakeup-source" will trigger the feature "RTC_WAKEUP_ALARM" which will
be send to the callers of the framework "rtc_get_info" PTA.
"wakeup-parent" permits to register locally the RTC's wakeup-parent, in
this case the exti.

- We configure our PM callback.

We also register the "stm32_rtc_alarm_wake_set_status" callback.
This callback should be called when the callers knows that the platform
will go in sleep mode. This setup the flag "alarm_wake" and in our PM
callback, if this one is set we enable the exti interrupt line that
will wakeup our platform.

Change-Id: I69faa7cc3e73de911e0f77aca950add21f99420e
Signed-off-by: Clément Le Goffic <clement.legoffic@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/402364
Domain-Review: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>