drivers: stm32_remoteproc: add API to get and release memory access
To be able to access to the memory to load the firmware images or
to clean the memories, it is necessary first to request access to
the firewall.
Add get and release ops to allow PTA to request the access to
the memories.
For the stm32mp1 the service is ignored and will be managed in a
separate commit.

Change-Id: If6ffc6ce09b1dc162ad428c65eb7a0c445333945
Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/391544
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Domain-Review: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>