regulator: add new ops init to configure the driver constraints
Add a new ops init which permit at the regulator drivers to manage a
default configuration of regulator parameters before to read the DT
properties.

The regulator properties are not mandatory in device tree, they can be
present to override the regulator default constraint value associated
with compatible when other constraint are present, board depending for
example.

This new ops allows to simplify the device tree and is aligned with
the Linux binding for regulators.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I8c50aee8952d68ff62e795cad67f18ad78eaf6eb
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/372635
Reviewed-by: Pascal PAILLET <p.paillet@st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>