drivers: stm32_iwdg: add stop watchdog handler
Implement .stop watchdog handler in order to allow the platform to
`halt` well.

The platform does not allow a way to stop an IWDG once started but a
solution exists.
We disable the platform reset ability of the IWDG we use and the IRQ
associated if the reset property is present in the DT.

Be careful, on STM32MP platforms, in A35TDCID, the system reset of the
IWDG1 cannot be disabled. Therefore, only the IWDG2 can be stopped.
The same behavior applies for IWDG3 in M33TDCID, you'll have to use the
IWDG4.

Change-Id: I23138c723052824380d7354ed3b6043df23f9836
Signed-off-by: Clément Le Goffic <clement.legoffic@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/381299
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@foss.st.com>
Domain-Review: Antonio Maria BORNEO <antonio.borneo@foss.st.com>