drivers: stm32mp_provisioning is dedicated to development
Explicitly state, in the source file and in a trace message
that use of stm32mp_provisioning is dedicated to development
support and not for an secure embedded product. When insecure config
support is disabled (i.e. CFG_WARN_INSECURE=n) explicitly panic if
some BSEC shadow cells are provisioned with this feature.

For consistency remove the panic instruction when there is no
embedded DTB and remove the "Warning:" wording from the debug trace
printed if the node has no cells.

Change-Id: I1d9d6a9d95d755e1d0fe034f4ccfbaf88a6e4973
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/382373
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Tested-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Domain-Review: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>