drivers: stm32_cpu_opp: skip OPP with unsupported frequency
Skip the OPP, if the associated frequency is not supported.

This patch allows to simplified the SoC and board device tree, as the
the list of OPP will be always present at SoC level but the list of
supported frequency for CPU are listed in RCC nodes "st,ck_mpu" and
sub-node of "st,clk_opp".

This behavior solves issue when a new OPP is added in SoC level
without modification on existing board.

Change-Id: I6f9ac3f6c37705c4a9cd9c84552364439d9807e4
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/382034
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>