drivers: stm32_uart: configure UART baudrate register
Add configuration of UART baudrate register in a UART driver probe. The
configuration of the baudrate register is done only for the UART device
used as serial console and once the UART is probed it is registered as
the serial console to use.

Change-Id: Iccfd1d7ed5bccb072dd0ba549a5d25dd5b84de82
Signed-off-by: Cheick TRAORE <cheick.traore-ext@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/374893
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Domain-Review: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>