drivers: stm32_rif: rework RIF device tree parsing
Rework the RIF device tree configuration parsing to rely on macros
defined in binding files. Also change the position of the resource ID
field in the RIFPROT macro so that it is present in the first 8 bits
instead of the 8 last.

Change-Id: Ia6ad4703af13fceccd2c83c902f8dc5f244f9ba1
Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/389467
Tested-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>