drivers: stm32_risab: configure SWRIAD after all other configs
If SRWIAD is disabled before the other configuration are done, there is
a gap where we can have secure writes forbidden to an area. This could
occur by cache evictions, especially from TF-A BL31. In opposition,
not enabling it if it is set in DT before configuration could lead to
the same kind of issues. Create 2 new functions enable_srwiad_if_set()
and disable_srwiad_if_unset(), that replace configure_srwiad(), and
will be called respectively before and after applying the rest of RISAB
configuration.

Change-Id: Idd520247990028c425c83ac0b44260b6f944ae1f
Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/392429
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>