scripts: sign_rproc_fw: Add support for AES CBC encryption
Allow the ELF firmware to be encrypted before signing.

In the first step, the ELF images are parsed and segments are encrypted
with the AES key provided in the "--enc_key" argument. This argument
also determines if the firmware needs to be encrypted.

In the case of encryption:
- The HASHTABLE TLV is replaced by the ENCTABLE TLV, which provides
  the hash of the encrypted segment and the associated random IV.
- The ENCTYPE TLV specifies the algorithm used for encryption.
- The IMGTYPE TLV is set with the new ENC_ELF type.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: Ib2109eb6fca7b38975949bee170e732c100d2a6c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/403797
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Domain-Review: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>