plat-stm32mp2: stm32mp25_pwr: avoid PM operation for Stop1 modes
The PM operation of PWR regulators to protect the IOs are needed only
for Standby modes and Stop2 modes; it is not needed for Stop1 modes
(as the associated regulators are not changed, only the D1 clocks are
stopped) and the duration of voltage switch of regulator can be important
with PMIC with ramp delay.

For Stop2 modes, the regulator supply for PWR IO domain can change,
so it is safe to protect the IO during low power modes.

For example, on STMicroelectronics boards the VDDIO1 for SDcard is
provided by STPMIC25 = VDDIO_SDCARD(LDO8) and is controlled by
pwrctrl3 = PWR_CPU_ON & NRST_FLASH_CPU (with st,pwrctrl-sel = <3> and
st,pwrctrl-reset). The supply is stopped in Stop2 and Standby1 low power
modes and restored tho NVM value on wake-up (regulator ON at 3V3).

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I16a99c47046dba7247b9ab43df821692a1a1e958
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/409248
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Tested-by: Christophe GUIBOUT <christophe.guibout@st.com>