crypto: stm32: fix SAES driver set_field_u32 usage
set_field_u32() is a function that allows you to change a specific bit
in a register by using a mask. The function returns the full value of
the register, which means that the use of bitwise OR here is a mistake.
The current code works here only because the modified registers are
initialized.

Fix the usage of the function in the SAES driver by replacing
bitwise OR assignments with simple assignments.

Signed-off-by: Maxime Méré <maxime.mere@foss.st.com>
Change-Id: I350d13e2d15ec9bd2d3483c0fe364f8d82988a5f
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/403658
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>