drivers: stm32_bsec3: allow driver file to be use without TDCID
If Cortex-A is not TDCID, it is still possible to use the BSEC mirror
filled by Cortex-M in SYSRAM. But the access to BSEC peripheral is
not possible in that case and will cause Illegal Accesses.
Protect those access by checking if the CFG_STM32_CM33TDCID is enabled
or not.

Change-Id: I2d5b36349bb111e9d8340d7b3db0e12f7df89c78
Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/409835
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>