drivers: add driver for DBGMCU secure debug mailbox for STM32MP21
On STM32MP21, the debug authentication protocol implements a bidirectional
communication between the debugger (host) and the target (STM32MP21)
through a mailbox interface located in the DBGMCU.
The host can write to the mailbox via the JTAG/SWD interface. It expects to
get responses and messages from the target via the same mailbox.

stm32_dbgmcu_mbx driver is introduced to demonstrate this mailbox and its
behaviors.
It requires to enable DBGMCU clock.
It requires two interrupts:
- dbg_auth_wr (write): it detects a debugger write in DBG_AUTH_HOST
register (request for debug of debugger that writes a key). It is cleared
when the CPU reads DBG_AUTH_HOST register or with POR.
- dbg_auth_rd (read): it detects a debugger read DBG_AUTH_DEVICE register
(debugger read response). It is also set with POR. It is cleared when the
CPU writes the DBG_AUTH_DEVICE register.

Change-Id: I52dd5411bcd56d0316e180cdce39c99b4cde07a0
Signed-off-by: Amelie Delaunay <amelie.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/433588
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>