dts: stm32: move scmi-server configuration in a separate file
Move SCMI server (based on SCP) configuration for STMicroelectronics
boards in a separate file "stm32mp25-st-scmi-cfg.dtsi", because
this configuration is not SoC generic but board specific.

This files can be reused by others boards or copied and customized
for specific needs.

Moreover this file prepare the CM33TDCID support as SCMI server
configuration is not required CM33TDCID.

This patch also reorders the "domain-name" element in STM32MP21
device treee to allow comparition with STM32MP23/25 files.

Change-Id: Ia98b68005964c385df25896e53a92235a0b4b55d
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/429882
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>