dts: stm32: add generic interrupt bindings for PWR IRQ
To prepare for the rework of PWR drivers to use the interrupt
framework, add:
- interrupt-controller property to PWR nodes;
- interrupts-extended and wakeup-source properties to the client
  nodes of PWR;
- interrupts property to PWR node of stm32mp131.

These changes are not used yet by the drivers that rely on the old
properties still present in the nodes.
The old properties would be dropped when will become unused due to
the rework of the drivers.

Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>
Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: Ib4ac9ee259f6608ef05fc11b8a8d721800df081c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/392425
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/441911
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Domain-Review: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>