clk: clk-stm32mp25: fix HSI/HSE/MSI Kernel clocks
Management of kernel clocks for HSI, HSE, and MSI clocks
for flexgen clock source.

These kernel clocks have their own gates KERON to activate this clock
in low power modes: the oscillator is kept running during Stop modes but
the kernel clock is not provided to the peripherals. The clock is then
available immediately when the system exits one of the Stop modes,
or when a peripheral requests the kernel clock.

Fixes: d1704414043c ("clk: clk-stm32mp25: add HSI/HSE/MSI Kernel clocks")
Change-Id: Id364105c936a0e0c34f2fbcd163fadc594555565
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/436787
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>