plat-stm32mp2: stm32mp25_pwr: remove fixme in driver
Remove a fixme in PWR regulator driver as it is NOT a fix
but a choice: use static pre-alloc instead dynamic allocation
as it is possible in new regulator framework.

The current choice is aligned with stm32mp1_pwr.c to avoid small
allocation during probe and because the number of PWR regulator is fixed.

Fixes: d4bb9f98f4af ("plat-stm32mp2: Add PWR regulators driver")
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Ieb83eec179cef827a3a030100980c458ce9f791f
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/438196
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>