drivers: remoteproc: reinitialize rproc fields in stm32_rproc_clean
rproc->boot_addr and rproc->tzen are set when we apply TLV configuration
on firmware image loading, and reset when we stop the remote processor.

If the load fails, we do not reinitialize the values as rproc_stop() is
never called. When we start a new firmware, we return an error because
the values are set twice.

Adding the variable reset in stm32_rproc_clean ensures that we return to
an initial state ready for a new load.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: Ic1ecb6990aebadb9fca2b48858650bfc1a8e9859
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/443460
Reviewed-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
Tested-by: Arnaud POULIQUEN <arnaud.pouliquen@st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Domain-Review: Arnaud POULIQUEN <arnaud.pouliquen@st.com>