clk: stm32: fix management of set_rate and determine_rate
When .determine_rate handler is implemented, the parent clock
and the actual clock are always computed and are applied by
clk_set_rate_no_lock(). Therefore the clock must implement .set_rate
handler instead of relying on CLK_SET_RATE_PARENT flag.

For cpu1/mpu clock it is not the case, so add the required ops set_rate()
to check the parent clock is the expected one after the call of
determine_rate().

Moreover for clocks with determine_rate ops, the flags CLK_SET_RATE_PARENT
is not needed (only used when determine_rate is absent), so we removes
this flags it when is possible.

Change-Id: Iddf9c4df4a439afaa1084938396bf181976bde45
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/447772
Reviewed-by: Gabriel FERNANDEZ <gabriel.fernandez@foss.st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@foss.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>