drivers: stm32_risaf: fix clock support
The clock for RISAF is associated to each device, managed by IP drivers
located in secure world with OP-TEE driver or in normal world with Linux
driver.

The RISAF driver enable clock in probe, so the clock is expected critical
and never stopped even is not used.

It is not perfecty for low power when clock is managed by OP-TEE and
exposed by SCMI BUT it is not working when clock is managed by non secure,
the clock is deactivated by Linux clock driver, that cause SERC in pm
function for suspend request.

It is the case for RISAF5 and PCIE on STM32MP25 boards when PCIE
is enabled in OP-TEE but not used in Linux kernel.

This patch just come back to previous behavior as short term solution.

Fixes: 0d37a2bfcee7 ("drivers: stm32_risaf: alignment with upstream")
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: Ie42481064c41980fe6b9bfca17cd98534da85e16
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/448588
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>